backend-flask/requirements.txt
flask
flask-cors
firebase-admin
python-dotenv
pyjwt
requests
